+++ enums/Enums.ts
@@ -148,6 +148,7 @@ export const SeasonStart = Object.freeze({
   CH2_S6: 1615852801,
   CH2_S7: 1623110401,
   CH2_S8: 1631491201,
+  CH3_S1: 1638662401,
 });
 
 export const SeasonEnd = Object.freeze({
@@ -168,6 +169,7 @@ export const SeasonEnd = Object.freeze({
   CH2_S5: 1615852800,
   CH2_S6: 1623110400,
   CH2_S7: 1631491200,
+  CH2_S8: 1638662400,
 });
 
 export const STWLeadSynergy = Object.freeze({